refactor: referral code block
